move all api fixtures out
